Fix bugs
Former-commit-id: fece28bfed8573d595dc585ab8cf8f254404c423
Former-commit-id: 60b06a566414f05444bdcff02a651df88a0df3b1 [formerly a21bcf1c46dd995cd91a3b2faed5a62be0be9ed2] [formerly 66a07fb38882193cd2a880e58bc76feb0f79c99e [formerly 0d66b7b1b32ac07b08d871ec61ff1a9509c56eee]] [formerly 689e48143df665f408e73196461b8ddfcd23ec49 [formerly f032dfafe80055a90a55b4c55eb38efdf908f849] [formerly 47c1d3978dec96af5e03edc12d8960f9386c589d [formerly f14a9f288b2ec3d0036f0c684ab1d47afb9d3db4]]]
Former-commit-id: f51ae773b21691e90f5f389118c83f0a927f4485 [formerly 66cb171d0ec0ba5284cf2c351e99e813cbf4c779] [formerly a201162dac9b1d01d20c5ac130499f7a1adbf258 [formerly 597da2ac02ac0b2ad6a97cc1a7f9878321dae291]]
Former-commit-id: b02d42a60c2fb62a041fa7b3df0263ce654fe3c5 [formerly d93888024dee4fff1e12e0a018c72d36b77e4a3a]
Former-commit-id: 76003919ad5ab7fb290c168b7e178ad366bc0c72

+++ src/USER-MINICELL/compute_plasmid_copy.cpp
@@ -25,6 +25,8 @@
 #include "group.h"
 #include <stdio.h>
 
+#include "comm.h"
+
 using namespace LAMMPS_NS;
 
 /* ---------------------------------------------------------------------- */
@@ -107,16 +109,16 @@ void ComputePlasmidCopy::compute_vector()
   for (int i = 0; i < nlocal; i++) {
     if (mask[i] & groupbit) {
       for (int j = 0; j < size_vector; j++) {
-	if (cpflag[j] == 0 && (int)fix_plasmid->vprop[i] == cp[j])
+	if (cpflag[j] == 0 && static_cast<int>(fix_plasmid->vprop[i]) == cp[j])
 	  vector[j]++;
-	if (cpflag[j] == 1 && (int)fix_plasmid->vprop[i] >= cp[j])
+	if (cpflag[j] == 1 && static_cast<int>(fix_plasmid->vprop[i]) >= cp[j])
 	  vector[j]++;
-	if (cpflag[j] == 2 && (int)fix_plasmid->vprop[i] <= cp[j])
+	if (cpflag[j] == 2 && static_cast<int>(fix_plasmid->vprop[i]) <= cp[j])
 	  vector[j]++;
       }
     }
   }
 
-  MPI_Allreduce(MPI_IN_PLACE,vector,size_vector,MPI_INT,MPI_SUM,world);
+  MPI_Allreduce(MPI_IN_PLACE,vector,size_vector,MPI_DOUBLE,MPI_SUM,world);
 }
